refactor: Move `preprocess_impedance_data` to utils
